rename dir, change path
