+++ features/plant/CategoryIndicator.tsx
