+++ features/plant/PlantAvatar.tsx
