+++ features/plant/PlantCard.tsx
