+++ features/plant/PlantCareScale.tsx
