+++ features/plant/PlantInfoBar.tsx
