+++ package.json
@@ -20,6 +20,7 @@
     "@types/react-dom": "18.0.11",
     "eslint": "8.34.0",
     "eslint-config-next": "^13.4.9",
+    "fuse.js": "^7.0.0",
     "mongodb": "^5.6.0",
     "mongoose": "^7.3.1",
     "next": "^13.4.9",
